Spotify Account Push Project
This project contains a custom script that cooperates with a table I created in Google sheets.

The table contains the members of my Spotify family account. The script checks the tables every Monday in order to find if any of the subscribers own me money and changes the colours of the cells depending on the validity. If anybody is overdue, the script sends two emails - both for me and for the member who should do the payment.